+++ src/Raven.CodeAnalysis/SemanticModel.Binding.cs
@@ -2812,11 +2812,12 @@ public partial class SemanticModel
         var typeSymbol = GetDeclaredTypeSymbol(declaration);
         var declarationBinder = new ClassDeclarationBinder(parentBinder, typeSymbol, declaration);
         declarationBinder.EnsureTypeParameterConstraintTypesResolved(typeSymbol.TypeParameters);
+        var effectiveDeclaration = GetEffectiveNominalTypeDeclaration(declaration);
 
         var valueType = Compilation.GetSpecialType(SpecialType.System_ValueType);
-        var defaultBaseType = GetDefaultBaseTypeForNominalDeclaration(declaration, objectType, valueType);
+        var defaultBaseType = GetDefaultBaseTypeForNominalDeclaration(effectiveDeclaration, objectType, valueType);
         var defaultInterfaces = GetDefaultNominalInterfaces(typeSymbol);
-        var shape = declarationBinder.BindNominalTypeShape(declaration, defaultBaseType, defaultInterfaces);
+        var shape = declarationBinder.BindNominalTypeShape(effectiveDeclaration, defaultBaseType, defaultInterfaces);
         var baseTypeSymbol = shape.BaseType;
         var interfaceList = shape.Interfaces;
 
@@ -2837,6 +2838,25 @@ public partial class SemanticModel
             ResolveSealedHierarchyPermits(declaration, typeSymbol, declarationBinder);
 
         classBinders.Add((declaration, declarationBinder));
+    }
+
+    private TypeDeclarationSyntax GetEffectiveNominalTypeDeclaration(TypeDeclarationSyntax declaration)
+    {
+        var effectiveDeclaration = declaration;
+
+        foreach (var attribute in declaration.AttributeLists.SelectMany(static list => list.Attributes))
+        {
+            if (!attribute.IsMacroAttribute())
+                continue;
+
+            if (GetMacroExpansion(attribute)?.ReplacementDeclaration is TypeDeclarationSyntax replacement &&
+                IsNominalTypeDeclaration(replacement))
+            {
+                effectiveDeclaration = replacement;
+            }
+        }
+
+        return effectiveDeclaration;
     }
 
     internal bool TryEnsureTypeMemberSignaturesDeclared(SourceNamedTypeSymbol typeSymbol)